Add sorted list of coins using $ revenue
Something worth noting is that Block Time seems to be a string in some
conis and float64 in others, and this directly affects the
unmarshalling step, which in case the Block Time is set as string will
only unmarshall coins that have it set as such. Whereas having it set as
float64, unmarshalling works for all coins and the coins that have it as
string simply set it to 0.

Signed-off-by: Antonio Gutierrez <chibby0ne@gmail.com>